+++ Procfile
@@ -1,3 +1,3 @@
-web: gunicorn hydra.app:app_factory --worker-class aiohttp.GunicornWebWorker
-scheduler: hydra-crawl
-release: hydra init_db --index
+web: gunicorn udata_hydra.app:app_factory --worker-class aiohttp.GunicornWebWorker
+scheduler: udata-hydra-crawl
+release: udata-hydra init_db --index
